feat(install): 应用商店增加版本更新检测功能
- 新增版本比较函数用于检测应用是否有可用更新
- 在应用列表中显示可更新标签提醒用户
- 优化应用详情页面文本大小样式

fix(system): 修复数据库字段兼容性问题

- 将日志操作表中的请求时间字段从decimal改为double类型
- 简化系统公告读取表的唯一索引定义
- 更新system应用版本至0.2.0并添加相应更新日志

+++ app/System/Models/LogOperate.php
@@ -21,7 +21,7 @@ class LogOperate extends Model
         $table->text('request_method')->comment("请求方法");
         $table->string('request_url')->comment("请求链接");
         $table->json('request_params')->comment("请求参数")->nullable();
-        $table->decimal('request_time', 11, 5)->comment("请求时间");
+        $table->double('request_time')->comment("请求时间");
         $table->string('route_name')->comment("路由名")->nullable();
         $table->string('route_title')->comment("路由标题")->nullable();
         $table->string('client_ua')->comment("客户端ua")->nullable();
